Updating NamespaceAndOutputPaths to be internal. Plus moving tests to match where the source code has been moved to.
