+++ test/EntityFramework.Commands.Tests/Scaffolding/Internal/ReverseEngineeringGeneratorTests.cs
@@ -1,12 +1,10 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using System.IO;
-using Microsoft.Data.Entity.Scaffolding.Internal;
 using Xunit;
 
-namespace Microsoft.Data.Entity.Relational.Design.ReverseEngineering
+namespace Microsoft.Data.Entity.Scaffolding.Internal
 {
     public class ReverseEngineeringGeneratorTests
     {
